+++ src/main/java/com/aclab/dne/model/MetaMatiereId.java
@@ -12,7 +12,7 @@ import java.io.Serializable;
 public class MetaMatiereId implements Serializable {
 
     private Long idPromotion;
-    private Long idDiplome;
+    private int idDiplome;
     private Long idUe;
     private Long idMatiere;
 }
